backend/flask_api/requirements.txt
Flask==2.2.3
torch==2.5.0
torchvision==0.20.0
Pillow==10.0.0
